+++ src/lib_modules/core/connection.cpp
@@ -32,7 +32,7 @@ size_t ConnectOutputToInput(IOutput *prev, IInput *next, IProcessExecutor * cons
 	});
 }
 
-size_t ConnectModules(IModule *prev, size_t outputIdx, IModule *next, size_t inputIdx, IProcessExecutor &executor) {
+size_t ConnectModules(IModule *prev, int outputIdx, IModule *next, int inputIdx, IProcessExecutor &executor) {
 	auto output = prev->getOutput(outputIdx);
 	return ConnectOutputToInput(output, next->getInput(inputIdx), &executor);
 }
